Move MediaLoader types to common location as it is required outside the media loader component
