+++ src/app/components/MediaLoader/types.ts
@@ -1,4 +1,9 @@
 import { PageTypes, Services } from '#app/models/types/global';
+import {
+  MediaType,
+  OnDemandAudioBlock,
+  OnDemandTVBlock,
+} from '#app/models/types/media';
 import { OptimoImageBlock } from '#app/models/types/optimo';
 import { Translations } from '#app/models/types/translations';
 
@@ -74,8 +79,6 @@ export type ConfigBuilderReturnProps = {
   showAds: boolean;
 };
 
-export type MediaType = 'audio' | 'video';
-
 export type MediaInfo = {
   title: string;
   datetime?: string;
@@ -141,7 +144,7 @@ export type AresMediaMetadataBlock = {
     text: string;
     title: string;
     imageUrl: string;
-    format: 'audio' | 'video';
+    format: MediaType;
     embedding: boolean;
     versions: {
       versionId: string;
@@ -162,7 +165,7 @@ export type AresMediaMetadataBlock = {
 export type ClipMediaBlock = {
   type: 'clipMedia';
   model: {
-    type: 'audio' | 'video';
+    type: MediaType;
     images: {
       source: string;
       urlTemplate: string;
@@ -178,50 +181,6 @@ export type ClipMediaBlock = {
       isEmbeddingAllowed: boolean;
     };
   };
-};
-
-export type OnDemandMediaModel = {
-  id: string;
-  subType: 'episode';
-  format: 'Video' | 'Audio';
-  title: string;
-  synopses: {
-    short: string;
-    medium: string;
-  };
-  imageUrl: string;
-  embedding: boolean;
-  advertising: boolean;
-  versions: [
-    {
-      versionId: string;
-      types: string[];
-      duration: number;
-      durationISO8601: string;
-      warnings: Record<string, string>;
-      availableTerritories: {
-        uk: boolean;
-        nonUk: boolean;
-        world: boolean;
-      };
-      availableFrom: number;
-      availabilityStatus: string;
-    },
-  ];
-  availability: string;
-  smpKind: string;
-  episodeTitle: string;
-  type: MediaType;
-};
-
-export type OnDemandTVBlock = {
-  type: 'onDemandTV';
-  model: OnDemandMediaModel;
-};
-
-export type OnDemandAudioBlock = {
-  type: 'onDemandAudio';
-  model: OnDemandMediaModel;
 };
 
 export type MediaBlock =
